Move getMemoryAlreadyBeingRevoked to utility class
so that it can be used by the PerQueryMemoryRevokingScheduler
